requirements.txt
pillow==10.2.0
omegaconf==2.3.0
einops==0.7.0
hydra-core==1.3.2
loguru==0.7.2
gputil==1.4.0
gradio==4.22.0
nvidia-smi==0.1.3
pyhocon==0.3.60
plotly==5.18.0
matplotlib==3.8.4
opencv-python==4.9.0.80
open3d==0.18.0
trimesh==4.1.3
roma==1.4.4
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
imageio==2.33.1
tensorboard==2.16.2
tqdm==4.66.1
plyfile==1.0.3
natsort==8.4.0
huggingface-hub[torch]>=0.22